jest: bin-program and bin-utils test cases refactored
